amplify/backend/api/graphql/src/requirements.txt
strawberry-graphql[fastapi]==0.209.7
uvicorn[standard]==0.23.2
gunicorn==21.0.1
fastapi[all]==0.104.1
boto3==1.29.5